libc: Rename fs/ to vfs/
