+++ lib/modules/libc/src/vfs/fchdirop.c
