+++ lib/modules/libc/src/vfs/fchdiroptab.c
